packages/python-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ccfoundry-agent-kit"
version = "0.1.0"
description = "Open-source agent-side SDK for building CoChiper Foundry-compatible agents"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "pydantic>=2.7,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "pyyaml>=6.0,<7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
